More work on shadow map rendering
Former-commit-id: 4c6e4a10afa48d9199f2857a2278072f756b5627

+++ Source/BansheeUtility/Source/BsDataStream.cpp
@@ -249,7 +249,7 @@ namespace bs
 	MemoryDataStream::MemoryDataStream(size_t size)
 		: DataStream(READ | WRITE), mData(nullptr), mFreeOnClose(true)
 	{
-		mData = mPos = (UINT8*)bs_alloc(size);
+		mData = mPos = (UINT8*)bs_alloc((UINT32)size);
 		mSize = size;
 		mEnd = mData + mSize;
